cleaned up repo, change video paths in tests
