+++ scripts/train_hydra.py
@@ -9,11 +9,13 @@ from pose_est_nets.models.regression_tracker import (
     RegressionTracker,
     SemiSupervisedRegressionTracker,
 )
-from pose_est_nets.models.new_heatmap_tracker import (
+from pose_est_nets.models.heatmap_tracker import (
     HeatmapTracker,
     SemiSupervisedHeatmapTracker,
 )
-from pose_est_nets.callbacks.freeze_unfreeze_callback import FeatureExtractorFreezeUnfreeze
+from pose_est_nets.callbacks.freeze_unfreeze_callback import (
+    FeatureExtractorFreezeUnfreeze,
+)
 from pytorch_lightning.loggers import TensorBoardLogger
 
 import os
@@ -110,18 +112,20 @@ def train(cfg: DictConfig):
             )
     logger = TensorBoardLogger("tb_logs", name="my_test_model")
     early_stopping = pl.callbacks.EarlyStopping(
-    monitor="val_loss", patience=100, mode="min"
+        monitor="val_loss", patience=100, mode="min"
     )
     lr_monitor = pl.callbacks.LearningRateMonitor(logging_interval="epoch")
 
-    ckpt_callback = pl.callbacks.model_checkpoint.ModelCheckpoint(monitor="val_loss")  
-    transfer_unfreeze_callback = FeatureExtractorFreezeUnfreeze(cfg.training.unfreezing_epoch) #Not used for now
+    ckpt_callback = pl.callbacks.model_checkpoint.ModelCheckpoint(monitor="val_loss")
+    transfer_unfreeze_callback = FeatureExtractorFreezeUnfreeze(
+        cfg.training.unfreezing_epoch
+    )  # Not used for now
     # TODO: add backbone refinement, add wandb?
     trainer = pl.Trainer(  # TODO: be careful with the devices here if you want to scale to multiple gpus
         gpus=1 if _TORCH_DEVICE == "cuda" else 0,
         max_epochs=cfg.training.max_epochs,
         log_every_n_steps=cfg.training.log_every_n_steps,
-        callbacks = [early_stopping, lr_monitor, ckpt_callback],
+        callbacks=[early_stopping, lr_monitor, ckpt_callback],
         logger=logger,
     )
     trainer.fit(model=model, datamodule=datamod)
@@ -129,4 +133,3 @@ def train(cfg: DictConfig):
 
 if __name__ == "__main__":
     train()  # I think you get issues when you try to get return values from a hydra function
-
